frontends/desktop-app/requirements.txt
PyQt5==5.15.7
python-dotenv==1.0.0
requests==2.31.0
